Refactored filesystem-dependent code to support.py
The idea is to hide the fact that we're using a filesystem (cf, eg, a
database) behind the support.py interface.

Actions which have not yet been updated in the refactored web_api.py have
also not been updated in mpt_cgi.py as yet.
